[Issue 2554] clean up fetcher frontend data fetcher pattern (#2567)
* removes the "services/search/searchfetcher" directory and existing "fetcher" pattern
* moves any necessary related functionality into the "/api" classes
