+++ frontend/src/app/api/SearchOpportunityAPI.ts
@@ -1,10 +1,10 @@
 import "server-only";
 
 import BaseApi from "src/app/api/BaseApi";
-import { QueryParamData } from "src/services/search/searchfetcher/SearchFetcher";
 import {
   PaginationOrderBy,
   PaginationRequestBody,
+  QueryParamData,
   SearchFetcherActionType,
   SearchFilterRequestBody,
   SearchRequestBody,
@@ -62,6 +62,12 @@ export default class SearchOpportunityAPI extends BaseApi {
       requestBody,
     );
 
+    response.actionType = searchInputs.actionType;
+    response.fieldChanged = searchInputs.fieldChanged;
+
+    if (!response.data) {
+      throw new Error("No data returned from Opportunity Search API");
+    }
     return response;
   }
 }
