+++ frontend/src/components/search/SearchPaginationFetch.tsx
@@ -1,7 +1,7 @@
 "use server";
 
-import { QueryParamData } from "src/services/search/searchfetcher/SearchFetcher";
-import { getSearchFetcher } from "src/services/search/searchfetcher/SearchFetcherUtil";
+import fetchers from "src/app/api/Fetchers";
+import { QueryParamData } from "src/types/search/searchRequestTypes";
 
 import SearchPagination from "src/components/search/SearchPagination";
 
@@ -16,8 +16,8 @@ export default async function SearchPaginationFetch({
   searchParams,
   scroll,
 }: SearchPaginationProps) {
-  const searchFetcher = getSearchFetcher();
-  const searchResults = await searchFetcher.fetchOpportunities(searchParams);
+  const searchResults =
+    await fetchers.searchOpportunityFetcher.searchOpportunities(searchParams);
   const totalPages = searchResults.pagination_info?.total_pages;
   const totalResults = searchResults.pagination_info?.total_records;
