+++ frontend/src/types/search/searchRequestTypes.ts
@@ -32,3 +32,16 @@ export enum SearchFetcherActionType {
   InitialLoad = "initialLoad",
   Update = "update",
 }
+
+export interface QueryParamData {
+  page: number;
+  query: string | null | undefined;
+  status: Set<string>;
+  fundingInstrument: Set<string>;
+  eligibility: Set<string>;
+  agency: Set<string>;
+  category: Set<string>;
+  sortby: string | null;
+  actionType?: SearchFetcherActionType;
+  fieldChanged?: string;
+}
